+++ srcs/ft_vector.c
@@ -8,7 +8,7 @@ t_vec	ft_create_vec(float x, float y, float z)
 
 t_vec	ft_normalize(t_vec v)
 {
-	double	magnitude;
+	float	magnitude;
 
 	magnitude = sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
 	v.x /= magnitude;
